requirements.txt
nltk
pandas
PyYAML
scikit-learn
torch==1.8.0
torchtext==0.9.0
tqdm